refactor(researcher): reorganize components into logical subdirectories
Move 25 flat component files into domain-based folders:
- chat/ (12 files): message rendering, input, chat container
- sources/ (3 files): source card, upload, my-stuff panel
- sidebar/ (5 files): navigation, history, user nav
- data/ (2 files): stream provider and handler
- auth/ (3 files): auth form, submit button, sign out

Also fix 3 pre-existing type errors:
- next-auth route handler signature for Next.js 16
- unpdf extractText unreachable branch with mergePages
- remove dead tool-getWeather code from message renderer
