+++ apps/researcher/app/(auth)/api/auth/[...nextauth]/route.ts
@@ -1,1 +1,4 @@
-export { GET, POST } from "@/app/(auth)/auth";
+import { GET as AuthGET, POST as AuthPOST } from "@/app/(auth)/auth";
+
+export const GET = AuthGET as (...args: any[]) => any;
+export const POST = AuthPOST as (...args: any[]) => any;
