+++ apps/researcher/app/(auth)/login/page.tsx
@@ -5,8 +5,8 @@ import { useRouter } from "next/navigation";
 import { useSession } from "next-auth/react";
 import { useActionState, useEffect, useState } from "react";
 
-import { AuthForm } from "@/components/auth-form";
-import { SubmitButton } from "@/components/submit-button";
+import { AuthForm } from "@/components/auth/auth-form";
+import { SubmitButton } from "@/components/auth/submit-button";
 import { toast } from "@/components/toast";
 import { type LoginActionState, login } from "../actions";
