+++ apps/researcher/app/(auth)/register/page.tsx
@@ -4,8 +4,8 @@ import Link from "next/link";
 import { useRouter } from "next/navigation";
 import { useSession } from "next-auth/react";
 import { useActionState, useEffect, useState } from "react";
-import { AuthForm } from "@/components/auth-form";
-import { SubmitButton } from "@/components/submit-button";
+import { AuthForm } from "@/components/auth/auth-form";
+import { SubmitButton } from "@/components/auth/submit-button";
 import { toast } from "@/components/toast";
 import { type RegisterActionState, register } from "../actions";
