+++ apps/researcher/app/(chat)/layout.tsx
@@ -1,8 +1,8 @@
 import { cookies } from "next/headers";
 import Script from "next/script";
 import { Suspense } from "react";
-import { AppSidebar } from "@/components/app-sidebar";
-import { DataStreamProvider } from "@/components/data-stream-provider";
+import { AppSidebar } from "@/components/sidebar/app-sidebar";
+import { DataStreamProvider } from "@/components/data/data-stream-provider";
 import { SidebarInset, SidebarProvider } from "@/components/ui/sidebar";
 import { auth } from "../(auth)/auth";
