+++ apps/researcher/app/(chat)/page.tsx
@@ -1,7 +1,7 @@
 import { cookies } from "next/headers";
 import { Suspense } from "react";
-import { Chat } from "@/components/chat";
-import { DataStreamHandler } from "@/components/data-stream-handler";
+import { Chat } from "@/components/chat/chat";
+import { DataStreamHandler } from "@/components/data/data-stream-handler";
 import { DEFAULT_CHAT_MODEL } from "@/lib/ai/models";
 import { generateUUID } from "@/lib/utils";
