+++ apps/researcher/components/auth/auth-form.tsx
@@ -1,7 +1,7 @@
 import Form from "next/form";
 
-import { Input } from "./ui/input";
-import { Label } from "./ui/label";
+import { Input } from "../ui/input";
+import { Label } from "../ui/label";
 
 export function AuthForm({
   action,
