+++ apps/researcher/components/auth/submit-button.tsx
@@ -4,7 +4,7 @@ import { useFormStatus } from "react-dom";
 
 import { LoaderIcon } from "@/components/icons";
 
-import { Button } from "./ui/button";
+import { Button } from "../ui/button";
 
 export function SubmitButton({
   children,
