+++ apps/researcher/components/chat/greeting.tsx
