+++ apps/researcher/components/chat/messages.tsx
@@ -2,7 +2,7 @@ import type { UseChatHelpers } from "@ai-sdk/react";
 import { ArrowDownIcon } from "lucide-react";
 import { useMessages } from "@/hooks/use-messages";
 import type { ChatMessage } from "@/lib/types";
-import { useDataStream } from "./data-stream-provider";
+import { useDataStream } from "../data/data-stream-provider";
 import { Greeting } from "./greeting";
 import { PreviewMessage, ThinkingMessage } from "./message";
