+++ apps/researcher/components/chat/suggested-actions.tsx
@@ -4,7 +4,7 @@ import type { UseChatHelpers } from "@ai-sdk/react";
 import { motion } from "framer-motion";
 import { memo } from "react";
 import type { ChatMessage } from "@/lib/types";
-import { Suggestion } from "./elements/suggestion";
+import { Suggestion } from "../elements/suggestion";
 import type { VisibilityType } from "./visibility-selector";
 
 type SuggestedActionsProps = {
