+++ apps/researcher/components/sidebar/sidebar-history-item.tsx
@@ -9,7 +9,7 @@ import {
   MoreHorizontalIcon,
   ShareIcon,
   TrashIcon,
-} from "./icons";
+} from "../icons";
 import {
   DropdownMenu,
   DropdownMenuContent,
@@ -19,12 +19,12 @@ import {
   DropdownMenuSubContent,
   DropdownMenuSubTrigger,
   DropdownMenuTrigger,
-} from "./ui/dropdown-menu";
+} from "../ui/dropdown-menu";
 import {
   SidebarMenuAction,
   SidebarMenuButton,
   SidebarMenuItem,
-} from "./ui/sidebar";
+} from "../ui/sidebar";
 
 const PureChatItem = ({
   chat,
