+++ apps/researcher/components/sidebar/sidebar-history.tsx
@@ -25,7 +25,7 @@ import {
 } from "@/components/ui/sidebar";
 import type { Chat } from "@/lib/db/schema";
 import { fetcher } from "@/lib/utils";
-import { LoaderIcon } from "./icons";
+import { LoaderIcon } from "../icons";
 import { ChatItem } from "./sidebar-history-item";
 
 type GroupedChats = {
