+++ apps/researcher/components/sidebar/sidebar-toggle.tsx
@@ -7,8 +7,8 @@ import {
   TooltipTrigger,
 } from "@/components/ui/tooltip";
 import { cn } from "@/lib/utils";
-import { SidebarLeftIcon } from "./icons";
-import { Button } from "./ui/button";
+import { SidebarLeftIcon } from "../icons";
+import { Button } from "../ui/button";
 
 export function SidebarToggle({
   className,
