+++ apps/researcher/components/sources/my-stuff-panel.tsx
