+++ apps/researcher/package.json
@@ -111,8 +111,8 @@
         "@types/d3-scale": "^4.0.8",
         "@types/node": "^22.8.6",
         "@types/papaparse": "^5.3.15",
-        "@types/react": "^18",
-        "@types/react-dom": "^18",
+        "@types/react": "^19",
+        "@types/react-dom": "^19",
         "@types/react-syntax-highlighter": "^15.5.13",
         "drizzle-kit": "^0.25.0",
         "postcss": "^8",
